+++ Experience/Experience.js
@@ -13,6 +13,7 @@ import World from './World/World';
 import Camera from './Camera';
 import Renderer from './Renderer';
 import Preloader from './Preloader';
+import Controls from './Controls';
 
 export default class Experience {
     static instance;
@@ -30,6 +31,10 @@ export default class Experience {
         this.resources = new Resources(assets);
         this.world = new World(); // order matters; only call after everything loads for World
         this.preloader = new Preloader();
+
+        this.preloader.on('enablecontrols', () => {
+            this.controls = new Controls();
+        });
 
         // Updates other classes' update functions when SIZE updates
         this.sizes.on('resize', () => {
