+++ Experience/Preloader.js
@@ -37,6 +37,7 @@ export default class Preloader extends EventEmitter {
         return new Promise((resolve) => {
             this.firstTimeline = new GSAP.timeline();
 
+            //===ANIMATE PRELOADER CUP===
             // Different preloader animations for desktop/mobile
             if (this.device === 'desktop') {
                 this.firstTimeline
@@ -57,6 +58,7 @@ export default class Preloader extends EventEmitter {
             } else {
                 // Device is mobile
                 this.roomChildren.cup_for_intro.position.set(0, -0.55, 0);
+                this.camera.orthographicCamera.rotation.set(-Math.PI / 7, 0, 0);
                 this.firstTimeline
                     .to(this.roomChildren.cup_for_intro.scale, {
                         x: 7,
@@ -79,322 +81,308 @@ export default class Preloader extends EventEmitter {
     // Move preloader back to center (will then be replaced by room model)
     secondIntro() {
         return new Promise((resolve) => {
+            //===REPLACE PRELOADER WITH ROOM MODEL===
+            // Same preloader animations for desktop/mobile
             this.secondTimeline = new GSAP.timeline();
 
-            // Different preloader animations for desktop/mobile
-            if (this.device === 'desktop') {
-                this.secondTimeline
-                    //===REPLACE PRELOADER WITH ROOM MODEL===
-
-                    //---Spin, scale, and center preloader cup---
-                    // Reset orthographic camera rotation
-                    .to(
-                        this.camera.orthographicCamera.rotation,
-                        {
-                            x: -Math.PI / 7
-                        },
-                        'spin_cup_room'
-                    )
-                    // Center preloader cup AND room
-                    .to(
-                        this.room.position,
-                        {
-                            x: 0,
-                            y: 0,
-                            z: 0,
-                            ease: 'power1.out'
-                        },
-                        'spin_cup_room'
-                    )
-                    // Spin preloader cup while centering
-                    .to(
-                        this.roomChildren.cup_for_intro.rotation,
-                        {
-                            z: Math.PI,
-                            ease: 'Cubic.InOut'
-                        },
-                        'spin_cup_room'
-                    )
-
-                    //---Enlarge preloader cup---
-                    //Scale preloader cup to room size
-                    .to(
-                        this.roomChildren.cup_for_intro.scale,
-                        {
-                            x: 44,
-                            y: 44,
-                            z: 27
-                        },
-                        'scale_cup_bigger'
-                    )
-                    // Move cup up on screen
-                    .to(
-                        this.roomChildren.cup_for_intro.position,
-                        {
-                            y: 1
-                        },
-                        'scale_cup_bigger'
-                    )
-
-                    //---Preloader cup disappears while room appears---
-                    // Unhide room model walls & floor
-                    .to(
-                        this.roomChildren.room_window.scale,
-                        {
-                            x: 1,
-                            y: 1,
-                            z: 1,
-                            duration: 0.7
-                        },
-                        'scale_cup_room_together'
-                    )
-                    // Spin room model walls & floor while appearing
-                    .to(
-                        this.room.rotation,
-                        {
-                            y: 4 * Math.PI // note: do we have to change rotation; not centered?
-                        },
-                        'scale_cup_room_together'
-                    )
-                    // Hide preloader cup
-                    .to(
-                        this.roomChildren.cup_for_intro.scale,
-                        {
-                            x: 0,
-                            y: 0,
-                            z: 0,
-                            duration: 1
-                        },
-                        'scale_cup_room_together'
-                    )
-                    // Spin preloader cup while disappearing
-                    .to(
-                        this.roomChildren.cup_for_intro.rotation,
-                        {
-                            z: 3 * Math.PI,
-                            ease: 'quartic.out'
-                        },
-                        'scale_cup_room_together'
-                    )
-                    // Move preloader cup up on the screen
-                    .to(
-                        this.roomChildren.cup_for_intro.position,
-                        {
-                            y: 0.2,
-                            ease: 'circular.out',
-                            duration: 0.3
-                        },
-                        'scale_cup_room_together'
-                    );
-
-                //===UNHIDE ROOM OBJECTS===
-                let shorten; // shorten declaration
-
-                // Unhide group desk
-                shorten = this.roomChildren.desk.children;
-                for (const ind in shorten) {
-                    this.secondTimeline.to(shorten[ind].scale, {
-                        x: this.roomChildrenScale[shorten[ind].name][0],
-                        y: this.roomChildrenScale[shorten[ind].name][1],
-                        z: this.roomChildrenScale[shorten[ind].name][2],
-                        ease: 'back.out(2.2)',
-                        duration: 0.15
-                    });
-                }
-
-                // Unhide group desk_stuff
-                shorten = this.roomChildren.desk_stuff.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(shorten[ind].scale, {
-                        x: this.roomChildrenScale[shorten[ind].name][0],
-                        y: this.roomChildrenScale[shorten[ind].name][1],
-                        z: this.roomChildrenScale[shorten[ind].name][2],
-                        ease: 'back.out(2.2)',
-                        duration: 0.2
-                    });
-                }
-
-                // Unhide group bag_pets + fish + memoboard
-                shorten = this.roomChildren.bag_pets.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<75%'
-                    );
-                }
-                shorten = this.roomChildren.fish; // shorten declaration
-                this.secondTimeline.to(
-                    shorten.scale,
+            //---Spin, scale, center preloader cup---
+            this.secondTimeline
+                // Reset orthographic camera rotation
+                .to(
+                    this.camera.orthographicCamera.rotation,
                     {
-                        x: this.roomChildrenScale[shorten.name][0],
-                        y: this.roomChildrenScale[shorten.name][1],
-                        z: this.roomChildrenScale[shorten.name][2],
-                        ease: 'back.out(2.2)',
-                        duration: 0.3
+                        x: -Math.PI / 7
                     },
-                    '<25%'
-                );
-                shorten = this.roomChildren.memoboard.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-
-                // Unhide group drawer
-                shorten = this.roomChildren.drawer.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-
-                // Unhide group exercise_whiteboard
-                shorten = this.roomChildren.exercise_whiteboard.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-
-                // Unhide group floor_stuff + painting
-                shorten = this.roomChildren.floor_stuff.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-                shorten = this.roomChildren.painting; // shorten declaration
-                this.secondTimeline.to(
-                    shorten.scale,
+                    'spin_cup_room'
+                )
+                // Center preloader cup AND room
+                .to(
+                    this.room.position,
                     {
-                        x: this.roomChildrenScale[shorten.name][0],
-                        y: this.roomChildrenScale[shorten.name][1],
-                        z: this.roomChildrenScale[shorten.name][2],
-                        ease: 'back.out(2.2)',
-                        duration: 0.3
-                    },
-                    '<25%'
-                );
-
-                // Unhide group lamp
-                shorten = this.roomChildren.lamp.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-
-                // Unhide group wall_shelf
-                shorten = this.roomChildren.wall_shelf.children; // shorten declaration
-                for (const ind in shorten) {
-                    this.secondTimeline.to(
-                        shorten[ind].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[ind].name][0],
-                            y: this.roomChildrenScale[shorten[ind].name][1],
-                            z: this.roomChildrenScale[shorten[ind].name][2],
-                            ease: 'back.out(2.2)',
-                            duration: 0.3
-                        },
-                        '<25%'
-                    );
-                }
-
-                // Unhide group chair
-                shorten = this.roomChildren.chair.children; // shorten declaration
-                // Unhide chair legs
-                this.secondTimeline.to(
-                    shorten[0].scale,
-                    {
-                        x: this.roomChildrenScale[shorten[0].name][0],
-                        y: this.roomChildrenScale[shorten[0].name][1],
-                        z: this.roomChildrenScale[shorten[0].name][2],
-                        ease: 'back.out(2.2)',
-                        duration: 0.3
-                    },
-                    '<25%'
-                );
-
-                // Unhide chair seat and spin
-                this.secondTimeline
-                    .to(
-                        shorten[1].scale,
-                        {
-                            x: this.roomChildrenScale[shorten[1].name][0],
-                            y: this.roomChildrenScale[shorten[1].name][1],
-                            z: this.roomChildrenScale[shorten[1].name][2],
-                            ease: 'power2.out',
-                            duration: 0.3
-                        },
-                        'chair_seat'
-                    )
-                    .to(
-                        shorten[1].rotation,
-                        {
-                            z: 2 * Math.PI
-                        },
-                        '<25%'
-                    );
-            } else {
-                // Device is mobile
-                this.roomChildren.cup_for_intro.position.set(0, -0.55, 0);
-                this.secondTimeline
-                    // Move preloader cup up
-                    .to(this.room.position, {
                         x: 0,
-                        z: 0,
                         y: 0,
-                        ease: 'back.out(2.5)',
+                        z: 0,
+                        ease: 'power1.out'
+                    },
+                    'spin_cup_room'
+                )
+                // Spin preloader cup while centering
+                .to(
+                    this.roomChildren.cup_for_intro.rotation,
+                    {
+                        z: Math.PI,
+                        ease: 'Cubic.InOut'
+                    },
+                    'spin_cup_room'
+                )
+
+                //---Enlarge preloader cup---
+                //Scale preloader cup to room size
+                .to(
+                    this.roomChildren.cup_for_intro.scale,
+                    {
+                        x: 44,
+                        y: 44,
+                        z: 27
+                    },
+                    'scale_cup_bigger'
+                )
+                // Move cup up on screen
+                .to(
+                    this.roomChildren.cup_for_intro.position,
+                    {
+                        y: 1
+                    },
+                    'scale_cup_bigger'
+                )
+
+                //---Preloader cup disappears while room appears---
+                // Unhide room model walls & floor
+                .to(
+                    this.roomChildren.room_window.scale,
+                    {
+                        x: 1,
+                        y: 1,
+                        z: 1,
                         duration: 0.7
-                    });
+                    },
+                    'scale_cup_room_together'
+                )
+                // Spin room model walls & floor while appearing
+                .to(
+                    this.room.rotation,
+                    {
+                        y: 4 * Math.PI // note: do we have to change rotation; not centered?
+                    },
+                    'scale_cup_room_together'
+                )
+                // Hide preloader cup
+                .to(
+                    this.roomChildren.cup_for_intro.scale,
+                    {
+                        x: 0,
+                        y: 0,
+                        z: 0,
+                        duration: 1
+                    },
+                    'scale_cup_room_together'
+                )
+                // Spin preloader cup while disappearing
+                .to(
+                    this.roomChildren.cup_for_intro.rotation,
+                    {
+                        z: 3 * Math.PI,
+                        ease: 'quartic.out'
+                    },
+                    'scale_cup_room_together'
+                )
+                // Move preloader cup up on the screen
+                .to(
+                    this.roomChildren.cup_for_intro.position,
+                    {
+                        y: 0.2,
+                        ease: 'circular.out',
+                        duration: 0.3
+                    },
+                    'scale_cup_room_together'
+                );
+
+            //===UNHIDE ROOM OBJECTS===
+            let shorten; // shorten declaration
+
+            // Unhide group desk
+            shorten = this.roomChildren.desk.children;
+            for (const ind in shorten) {
+                this.secondTimeline.to(shorten[ind].scale, {
+                    x: this.roomChildrenScale[shorten[ind].name][0],
+                    y: this.roomChildrenScale[shorten[ind].name][1],
+                    z: this.roomChildrenScale[shorten[ind].name][2],
+                    ease: 'back.out(2.2)',
+                    duration: 0.15
+                });
             }
+
+            // Unhide group desk_stuff
+            shorten = this.roomChildren.desk_stuff.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(shorten[ind].scale, {
+                    x: this.roomChildrenScale[shorten[ind].name][0],
+                    y: this.roomChildrenScale[shorten[ind].name][1],
+                    z: this.roomChildrenScale[shorten[ind].name][2],
+                    ease: 'back.out(2.2)',
+                    duration: 0.2
+                });
+            }
+
+            // Unhide group bag_pets + fish + memoboard
+            shorten = this.roomChildren.bag_pets.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<75%'
+                );
+            }
+            shorten = this.roomChildren.fish; // shorten declaration
+            this.secondTimeline.to(
+                shorten.scale,
+                {
+                    x: this.roomChildrenScale[shorten.name][0],
+                    y: this.roomChildrenScale[shorten.name][1],
+                    z: this.roomChildrenScale[shorten.name][2],
+                    ease: 'back.out(2.2)',
+                    duration: 0.3
+                },
+                '<25%'
+            );
+            shorten = this.roomChildren.memoboard.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+
+            // Unhide group drawer
+            shorten = this.roomChildren.drawer.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+
+            // Unhide group exercise_whiteboard
+            shorten = this.roomChildren.exercise_whiteboard.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+
+            // Unhide group floor_stuff + painting
+            shorten = this.roomChildren.floor_stuff.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+            shorten = this.roomChildren.painting; // shorten declaration
+            this.secondTimeline.to(
+                shorten.scale,
+                {
+                    x: this.roomChildrenScale[shorten.name][0],
+                    y: this.roomChildrenScale[shorten.name][1],
+                    z: this.roomChildrenScale[shorten.name][2],
+                    ease: 'back.out(2.2)',
+                    duration: 0.3
+                },
+                '<25%'
+            );
+
+            // Unhide group lamp
+            shorten = this.roomChildren.lamp.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+
+            // Unhide group wall_shelf
+            shorten = this.roomChildren.wall_shelf.children; // shorten declaration
+            for (const ind in shorten) {
+                this.secondTimeline.to(
+                    shorten[ind].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[ind].name][0],
+                        y: this.roomChildrenScale[shorten[ind].name][1],
+                        z: this.roomChildrenScale[shorten[ind].name][2],
+                        ease: 'back.out(2.2)',
+                        duration: 0.3
+                    },
+                    '<25%'
+                );
+            }
+
+            //---Unhide group chair---
+            shorten = this.roomChildren.chair.children; // shorten declaration
+            // Unhide chair legs
+            this.secondTimeline.to(
+                shorten[0].scale,
+                {
+                    x: this.roomChildrenScale[shorten[0].name][0],
+                    y: this.roomChildrenScale[shorten[0].name][1],
+                    z: this.roomChildrenScale[shorten[0].name][2],
+                    ease: 'back.out(2.2)',
+                    duration: 0.3
+                },
+                '<25%'
+            );
+
+            // Unhide chair seat and spin
+            this.secondTimeline
+                .to(
+                    shorten[1].scale,
+                    {
+                        x: this.roomChildrenScale[shorten[1].name][0],
+                        y: this.roomChildrenScale[shorten[1].name][1],
+                        z: this.roomChildrenScale[shorten[1].name][2],
+                        ease: 'power2.out',
+                        duration: 0.3
+                    },
+                    'chair_seat'
+                )
+                .to(
+                    shorten[1].rotation,
+                    {
+                        z: 2 * Math.PI,
+                        onComplete: resolve // signals end of animation
+                    },
+                    '<25%'
+                );
         });
     }
 
@@ -402,20 +390,49 @@ export default class Preloader extends EventEmitter {
     onScroll(e) {
         // If user scrolls down
         if (e.deltaY > 0) {
-            window.removeEventListener('wheel', this.scrollOnce);
+            this.removeEventListeners();
             this.playSecondIntro();
         }
+    }
+
+    // Get user's current touch
+    onTouch(e) {
+        this.initialY = e.touches[0].clientY;
+    }
+
+    // Move screen if user swipes up
+    onTouchMove(e) {
+        let currentY = e.touches[0].clientY;
+        let diff = this.initialY - currentY;
+        // If user swipes up, play second intro
+        if (diff > 0) {
+            this.removeEventListeners();
+            this.playSecondIntro();
+        }
+        // Resets user's touch
+        this.initialY = null;
+    }
+
+    removeEventListeners() {
+        window.removeEventListener('wheel', this.scrollOnce);
+        window.removeEventListener('touchstart', this.touchStart);
+        window.removeEventListener('touchmove', this.touchMove);
     }
 
     async playIntro() {
         // Wait until firstIntro is done before proceeding
         await this.firstIntro();
         this.scrollOnce = this.onScroll.bind(this); // allows instance of onScroll func to be removed
+        this.touchStart = this.onTouch.bind(this);
+        this.touchMove = this.onTouchMove.bind(this);
         window.addEventListener('wheel', this.scrollOnce);
+        window.addEventListener('touchstart', this.touchStart);
+        window.addEventListener('touchmove', this.touchMove);
     }
 
     async playSecondIntro() {
         // Wait until secondIntro is done before proceeding
         await this.secondIntro();
+        this.emit('enablecontrols');
     }
 }
