+++ style.css
@@ -13,9 +13,9 @@ body {
     color: var(--text-black);
 }
 
-body {
-    overflow: hidden; /* hide default scrollbar */
-}
+/* body {
+    overflow: hidden;
+} */
 
 :root {
     /* Use a color close to but not exactly black */
@@ -47,7 +47,7 @@ LANDING PAGE
 .page {
     z-index: 99999;
     width: 100%;
-    height: 100%;
+    height: 100vh; /* hide default scrollbar */
     overflow: hidden;
 }
